requirements.txt
hydra-core==1.3.2
omegaconf==2.3.0
sentencepiece==0.1.99
torch==2.0.1
transformers==4.41.1
wandb==0.15.8
protobuf==4.25.3
matplotlib
tqdm
scikit-learn
scipy
numpy